move nft wasm db into wasm module, add tables into wasm
